backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.10.3
python-dotenv==1.0.1

# LLM
anthropic>=0.50.0

# Observability / evaluation
langsmith>=0.1.147

# Content extraction
youtube-transcript-api>=0.6.3
trafilatura>=1.12.0
httpx>=0.27.0

# Web search
tavily-python>=0.3.9
